Narrow puzzle hints to tor ring and optional starter flats
