+++ js/gamemaker/shift-host.js
@@ -7,6 +7,7 @@ import { applyColumnShiftInPlace, applyRowShiftInPlace } from "../board-logic.js
  *   getGameBoard: () => string[][];
  *   syncDomFromBoard: () => void;
  *   syncLineOverlaySize: () => void;
+ *   scheduleSyncLineOverlaySize: () => void;
  *   lockGridSizeForSwipe: () => void;
  *   unlockGridSizeAfterSwipe: () => void;
  *   getIsGameActive: () => boolean;
@@ -19,6 +20,7 @@ export function createGamemakerShiftHost(deps) {
     getGameBoard,
     syncDomFromBoard,
     syncLineOverlaySize,
+    scheduleSyncLineOverlaySize,
     lockGridSizeForSwipe,
     unlockGridSizeAfterSwipe,
     getIsGameActive,
@@ -64,6 +66,7 @@ export function createGamemakerShiftHost(deps) {
       shiftHost.syncDomFromBoard();
     },
     syncLineOverlaySize,
+    scheduleSyncLineOverlaySize,
     clearTapStreak: () => {
       ctx.state.shift.doubleTapPrevAt = 0;
     },
