+++ package.json
@@ -8,7 +8,8 @@
     "gen:word-rec": "node scripts/export-word-recognizability.mjs",
     "gen:extend-metrics": "python3 scripts/build-extended-word-metrics.py",
     "gen:puzzle-pool": "node scripts/generate-puzzle-pool.mjs",
-    "gen:puzzle-wordlist": "node scripts/export-puzzle-wordlist-seed.mjs"
+    "gen:puzzle-wordlist": "node scripts/export-puzzle-wordlist-seed.mjs",
+    "inventory:hints": "node scripts/inventory-puzzle-hints.mjs"
   },
   "devDependencies": {
     "prettier": "3.1.0"
